pyproject.toml
[tool.poetry]
name = "delta"
version = "1.0"
description = "A multipurpouse discord bot."
authors = ["yorii260 <bordbox4@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
discord = "^2.3.2"
python-dotenv = "^1.0.0"
pymongo = "^4.5.0"
python-dateutil = "^2.8.2"
motor = "^3.3.1"
umongo = "^3.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
